switch to YAML config.
