--- config/fixtures/nonexistent_nc_name.conf.input
@@ -1,18 +1,0 @@
-notification_config {
-  name: "alertmanager_test"
-  pagerduty_config {
-    service_key: "supersecretapikey"
-  }
-  email_config {
-    email: "test@testservice.org"
-  }
-}
-
-aggregation_rule {
-  filter {
-    name_re: "service"
-    value_re: "test"
-  }
-  repeat_rate_seconds: 3600
-  notification_config_name: "alertmanager_test2"
-}
