--- config/fixtures/sample.conf.input
@@ -1,36 +1,0 @@
-notification_config {
-  name: "alertmanager_test"
-  pagerduty_config {
-    service_key: "supersecretapikey"
-  }
-  email_config {
-    email: "test@testservice.org"
-  }
-  pushover_config {
-    token: "mypushovertoken"
-    user_key: "mypushoverkey"
-  }
-  hipchat_config {
-    auth_token: "hipchatauthtoken"
-    room_id: 123456
-    send_resolved: true
-  }
-  slack_config {
-    webhook_url: "webhookurl"
-    send_resolved: true
-  }
-  flowdock_config {
-    api_token: "4c7234902348234902384234234cdb59"
-    from_address: "aliaswithgravatar@somehwere.com"
-    tag: "monitoring"
-  }
-}
-
-aggregation_rule {
-  filter {
-    name_re: "service"
-    value_re: "test"
-  }
-  repeat_rate_seconds: 3600
-  notification_config_name: "alertmanager_test"
-}
